feat: 模型事件钩子和自增自减方法优化
- 添加 ModelEvent trait 支持增删改查前后钩子
- creating/created - 创建前/后
- updating/updated - 更新前/后
- saving/saved - 保存前/后（插入/更新）
- deleting/deleted - 删除前/后
- restoring/restored - 恢复前/后（软删除）
- 修改 increment/decrement 为 inc/dec

+++ src/Model/Model.php
@@ -9,6 +9,7 @@ use Kode\Database\Model\Concerns\HasAttributes;
 use Kode\Database\Model\Concerns\SoftDeletes;
 use Kode\Database\Model\Concerns\Timestamps;
 use Kode\Database\Model\Concerns\QueriesRelationships;
+use Kode\Database\Model\ModelEvent;
 use ArrayAccess;
 use JsonSerializable;
 
@@ -24,6 +25,7 @@ abstract class Model implements ArrayAccess, JsonSerializable
     use Timestamps;
     use SoftDeletes;
     use QueriesRelationships;
+    use ModelEvent;
 
     protected string $table = '';
     protected string $primaryKey = 'id';
@@ -75,13 +77,23 @@ abstract class Model implements ArrayAccess, JsonSerializable
      */
     public function save(): bool
     {
+        if (!$this->beforeSave()) {
+            return false;
+        }
+
         $this->setTimestamps();
 
         if ($this->exists) {
-            return $this->performUpdate();
+            $result = $this->performUpdate();
+        } else {
+            $result = $this->performInsert();
         }
 
-        return $this->performInsert();
+        if ($result) {
+            $this->afterSave();
+        }
+
+        return $result;
     }
 
     /**
@@ -164,8 +176,15 @@ abstract class Model implements ArrayAccess, JsonSerializable
         }
 
         if ($this->usesSoftDeletes()) {
+            if (!$this->beforeDelete()) {
+                return false;
+            }
             $this->{$this->softDeleteField} = date($this->dateFormat);
-            return $this->save();
+            $result = $this->save();
+            if ($result) {
+                $this->afterDelete();
+            }
+            return $result;
         }
 
         return $this->forceDelete();
@@ -176,6 +195,10 @@ abstract class Model implements ArrayAccess, JsonSerializable
      */
     public function forceDelete(): bool
     {
+        if (!$this->beforeDelete()) {
+            return false;
+        }
+
         $sql = sprintf(
             'DELETE FROM %s WHERE %s = ?',
             $this->table,
@@ -186,6 +209,7 @@ abstract class Model implements ArrayAccess, JsonSerializable
 
         if ($affected > 0) {
             $this->exists = false;
+            $this->afterDelete();
         }
 
         return $affected > 0;
@@ -200,8 +224,18 @@ abstract class Model implements ArrayAccess, JsonSerializable
             return false;
         }
 
+        if (!$this->beforeRestore()) {
+            return false;
+        }
+
         $this->{$this->softDeleteField} = null;
-        return $this->save();
+        $result = $this->save();
+
+        if ($result) {
+            $this->afterRestore();
+        }
+
+        return $result;
     }
 
     /**
